requirements.txt
python-dotenv >= 1.0.0
watchdog >= 0.8.2
flask >= 3.0.0
cryptography >= 41.0.7
ghostscript >= 0.7
PyPDF2 >= 3.0.1